+++ src/library/DIPS.Mobile.UI/Components/Searching/SearchPage.cs
@@ -1,8 +1,3 @@
-#if __IOS__
-using DIPS.Mobile.UI.Components.Searching.iOS;
-using UIKit;
-#endif
-
 using DIPS.Mobile.UI.Components.Lists;
 using DIPS.Mobile.UI.Extensions;
 using DIPS.Mobile.UI.Internal;
@@ -21,11 +16,10 @@ namespace DIPS.Mobile.UI.Components.Searching
 
         private View? m_previousView;
         private View? m_footerView;
+        private View? m_headerView;
 
         public SearchPage()
         {
-            Loaded += OnLoaded;
-            Unloaded += OnUnLoaded;
             
             //Searchbar
             SearchBar = new SearchBar { AutomationId = "SearchBar".ToDUIAutomationId<SearchPage>(), HasCancelButton = true, HasBusyIndication = true, ShouldCloseKeyboardOnReturnKeyTapped = true };
@@ -75,7 +69,7 @@ namespace DIPS.Mobile.UI.Components.Searching
                 {
                     new()
                     {
-                        Height = GridLength.Auto // Space for safe area ios
+                        Height = GridLength.Auto // Space for optional header
                     },
                     new()
                     {
@@ -109,18 +103,13 @@ namespace DIPS.Mobile.UI.Components.Searching
             SearchBar.Unfocus();
         }
 
-        private void OnUnLoaded(object? sender, EventArgs e)
+        protected override void OnHandlerChanged()
         {
-            Loaded -= OnLoaded;
-            Unloaded -= OnUnLoaded;
-            m_resultCollectionView.Scrolled -= OnCollectionViewScrolled;
+            base.OnHandlerChanged();
             
-            SearchBar.Focused -= OnSearchBarFocused;
+            if (Handler is null)
+                return;
             
-        }
-
-        private void OnLoaded(object? sender, EventArgs e)
-        {
             SetSearchState(SearchStates.NeedsSearchHint);
             if (ShouldAutoFocus)
             {
@@ -128,22 +117,6 @@ namespace DIPS.Mobile.UI.Components.Searching
             }
 
             SearchBar.Focused += OnSearchBarFocused;
-
-#if __IOS__
-            if (OperatingSystem.IsIOSVersionAtLeast(14, 1))
-            {
-                var safeAreaInsetsTop = UIApplication.SharedApplication.KeyWindow.SafeAreaInsets.Top;
-                
-                var topBoxView = new BoxView
-                {
-                    BackgroundColor = SearchBar.BarColor, HeightRequest = safeAreaInsetsTop
-                };
-                
-                m_grid.Children.Insert(0, topBoxView);
-
-                topBoxView.Margin = new Thickness(0, -safeAreaInsetsTop, 0, 0);
-            }
-#endif
         }
 
         private void OnSearchBarFocused(object? sender, EventArgs eventArgs)
@@ -262,6 +235,17 @@ namespace DIPS.Mobile.UI.Components.Searching
             SearchBar.ReturnKeyType = SearchMode == SearchMode.WhenTextChanged ?
                 ReturnType.Done :
                 ReturnType.Search;
+        }
+
+        private void OnHeaderViewChanged()
+        {
+            if (m_headerView is not null)
+                m_grid.Remove(m_headerView);
+
+            m_headerView = HeaderView;
+            
+            if (m_headerView is not null)
+                m_grid.Add(m_headerView, 0, 0);
         }
 
         private void OnFooterViewChanged()